Refactoring of app.tsx, more modular
